feat: Add coordinates and registered_at field to work shifts
This commit adds the 'coordinates' field to the work shifts table, which stores the latitude and longitude of where the work shift was registered. Additionally, it adds the 'registered_at' field, which represents the timestamp when the work shift was registered on the mobile device.

Also made seeders adjusts for these new fields